+++ packages/react/src/Link.ts
@@ -53,6 +53,9 @@ const Link = forwardRef<unknown, InertiaLinkProps>(
       onCancel = noop,
       onSuccess = noop,
       onError = noop,
+      onHttpException = noop,
+      onNetworkError = noop,
+      onFlash = noop,
       onPrefetching = noop,
       onPrefetched = noop,
       prefetch = false,
@@ -154,6 +157,9 @@ const Link = forwardRef<unknown, InertiaLinkProps>(
         onCancel,
         onSuccess,
         onError,
+        onHttpException,
+        onNetworkError,
+        onFlash,
       }),
       [
         baseParams,
@@ -166,6 +172,9 @@ const Link = forwardRef<unknown, InertiaLinkProps>(
         onCancel,
         onSuccess,
         onError,
+        onHttpException,
+        onNetworkError,
+        onFlash,
       ],
     )
 
@@ -251,7 +260,11 @@ const Link = forwardRef<unknown, InertiaLinkProps>(
       onMouseLeave: () => {
         clearTimeout(hoverTimeout.current)
       },
-      onClick: regularEvents.onClick,
+      onClick: (event: React.MouseEvent) => {
+        clearTimeout(hoverTimeout.current)
+
+        regularEvents.onClick(event)
+      },
     }
 
     const prefetchClickEvents = {