+++ packages/svelte/src/createInertiaApp.ts
@@ -143,6 +143,7 @@ export default async function createInertiaApp<SharedProps extends PageProps = P
         resolveLoading,
         defaultLayout: layout,
         layer,
+        serverRendered: true,
       }
 
       let svelteApp: SvelteRenderResult
@@ -191,6 +192,8 @@ export default async function createInertiaApp<SharedProps extends PageProps = P
         )
       : null
 
+  const el = isServer ? null : document.getElementById(id)!
+
   const props: InertiaAppProps<SharedProps> = {
     initialPage,
     initialComponent,
@@ -199,6 +202,7 @@ export default async function createInertiaApp<SharedProps extends PageProps = P
     resolveLoading,
     defaultLayout: layout,
     layer,
+    serverRendered: isServer || el!.hasAttribute('data-server-rendered'),
   }
 
   // SSR with page provided (legacy pattern used by ssr.ts)
@@ -222,7 +226,8 @@ export default async function createInertiaApp<SharedProps extends PageProps = P
   }
 
   // CSR
-  const target = document.getElementById(id)!
+  const target = el!
+  const isServerRendered = props.serverRendered!
 
   if (setup) {
     await setup({ el: target, App, props })
@@ -233,7 +238,7 @@ export default async function createInertiaApp<SharedProps extends PageProps = P
       withApp(context, { ssr: false, page: initialPage })
     }
 
-    if (target.hasAttribute('data-server-rendered')) {
+    if (isServerRendered) {
       hydrate(App, { target, props, context })
     } else {
       mount(App, { target, props, context })