+++ packages/svelte/src/link.ts
@@ -35,6 +35,9 @@ type SelectedEventKeys =
   | 'cancel'
   | 'success'
   | 'error'
+  | 'httpException'
+  | 'networkError'
+  | 'flash'
   | 'prefetching'
   | 'prefetched'
 type SelectedGlobalEventsMap = Pick<GlobalEventsMap, SelectedEventKeys>
@@ -64,19 +67,25 @@ function link(
   let baseParams: VisitOptions
   let visitParams: VisitOptions
 
+  const handleClick = (event: MouseEvent) => {
+    if (shouldIntercept(event)) {
+      event.preventDefault()
+      router.visit(href, visitParams)
+    }
+  }
+
   const regularEvents: ActionEventHandlers = {
-    click: (event: MouseEvent) => {
-      if (shouldIntercept(event)) {
-        event.preventDefault()
-        router.visit(href, visitParams)
-      }
-    },
+    click: handleClick,
   }
 
   const prefetchHoverEvents: ActionEventHandlers = {
     mouseenter: () => (hoverTimeout = setTimeout(() => prefetch(), config.get('prefetch.hoverDelay'))),
     mouseleave: () => clearTimeout(hoverTimeout),
-    click: regularEvents.click,
+    click: (event: MouseEvent) => {
+      clearTimeout(hoverTimeout)
+
+      handleClick(event)
+    },
   }
 
   const prefetchClickEvents: ActionEventHandlers = {
@@ -206,6 +215,9 @@ function link(
       onCancel: () => dispatchEvent('cancel'),
       onSuccess: (page) => dispatchEvent('success', { detail: { page } }),
       onError: (errors) => dispatchEvent('error', { detail: { errors } }),
+      onHttpException: (response) => dispatchEvent('httpException', { cancelable: true, detail: { response } }),
+      onNetworkError: (error) => dispatchEvent('networkError', { cancelable: true, detail: { error } }),
+      onFlash: (flash) => dispatchEvent('flash', { detail: { flash } }),
       onCancelToken: (token) => dispatchEvent('cancel-token', { detail: { token } }),
       onPrefetching: (visit) => dispatchEvent('prefetching', { detail: { visit } }),
       onPrefetched: (response, visit) => dispatchEvent('prefetched', { detail: { response, visit } }),