+++ packages/vite/src/index.ts
@@ -54,8 +54,14 @@ export interface InertiaPluginOptions {
    *     extractDefault: true,
    *     ssr: (configureCall, options) => `
    *       import createServer from '@inertiajs/solid/server'
-   *       const render = await ${configureCall}
-   *       createServer((page) => render(page)${options})
+   *       const renderPromise = ${configureCall}
+   *       renderPromise.catch((error) => console.error(error))
+   *       const renderPage = async (page) => {
+   *         const render = await renderPromise
+   *
+   *         return render(page)
+   *       }
+   *       createServer(renderPage${options})
    *     `,
    *   }
    * })