+++ packages/vue3/src/index.ts
@@ -18,6 +18,7 @@ export { default as useLayer } from './useLayer'
 export { default as usePoll } from './usePoll'
 export { default as usePrefetch } from './usePrefetch'
 export { default as useRemember } from './useRemember'
+export { default as WhenMounted } from './whenMounted'
 export { default as WhenVisible } from './whenVisible'
 
 export const config = coreConfig.extend<VueInertiaAppConfig>({})